collector-api-flask/app/requirements.txt
python-dotenv
urllib3
pyjwt
pymongo
influxdb-client
flask
